rename modules, tree error variant
