+++ rust/routee-compass-core/src/model/traversal/default/custom/custom_traversal_service.rs
@@ -1,7 +1,7 @@
 use super::{CustomTraversalEngine, CustomTraversalModel};
 use crate::model::traversal::{
-    traversal_model::TraversalModel, traversal_model_error::TraversalModelError,
-    traversal_model_service::TraversalModelService,
+    model::TraversalModel, error::TraversalModelError,
+    service::TraversalModelService,
 };
 use std::sync::Arc;
