+++ rust/routee-compass-core/src/model/traversal/default/distance/distance_traversal_model.rs
@@ -7,8 +7,8 @@ use crate::model::state::StateModel;
 use crate::model::state::StateVariable;
 use crate::model::state::{InputFeature, StateVariableConfig};
 use crate::model::traversal::default::fieldname;
-use crate::model::traversal::traversal_model::TraversalModel;
-use crate::model::traversal::traversal_model_error::TraversalModelError;
+use crate::model::traversal::model::TraversalModel;
+use crate::model::traversal::error::TraversalModelError;
 use crate::model::unit::DistanceUnit;
 use crate::util::geo::haversine;
