+++ rust/routee-compass-core/src/model/traversal/mod.rs
@@ -1,14 +1,14 @@
 pub mod default;
 mod edge_traversal_context;
-mod traversal_model;
-mod traversal_model_builder;
-mod traversal_model_error;
-mod traversal_model_service;
-mod traversal_result;
+mod model;
+mod builder;
+mod error;
+mod service;
+mod result;
 
 pub use edge_traversal_context::EdgeFrontierContext;
-pub use traversal_model::TraversalModel;
-pub use traversal_model_builder::TraversalModelBuilder;
-pub use traversal_model_error::TraversalModelError;
-pub use traversal_model_service::TraversalModelService;
-pub use traversal_result::TraversalResult;
+pub use model::TraversalModel;
+pub use builder::TraversalModelBuilder;
+pub use error::TraversalModelError;
+pub use service::TraversalModelService;
+pub use result::TraversalResult;
