+++ rust/routee-compass-core/src/model/traversal/model.rs
@@ -1,4 +1,4 @@
-use super::traversal_model_error::TraversalModelError;
+use super::error::TraversalModelError;
 use crate::algorithm::search::SearchTree;
 use crate::model::network::Vertex;
 use crate::model::state::{InputFeature, StateModel, StateVariable, StateVariableConfig};
